+++ components/Skeleton/Skeleton.module.css
@@ -1,19 +1,17 @@
 .skeleton {
-  width: 100%;
-  background: linear-gradient(to right, #333 0%, #555 50%, #333 100%);
-  background-size: 400% 400%;
-  animation: pulse 1s ease-in-out infinite;
-  min-height: 12px;
-  max-height: 100%;
-  padding: 2px;
-  margin: 2px;
+	width: 100%;
+	background: linear-gradient(to right, #333 0%, #555 50%, #333 100%);
+	background-size: 400% 400%;
+	animation: pulse 1s ease-in-out infinite;
+	min-height: 12px;
+	max-height: 100%;
 }
 
 @keyframes pulse {
-  0% {
-    background-position: 0% 0%;
-  }
-  100% {
-    background-position: -135% 0%;
-  }
+	0% {
+		background-position: 0% 0%;
+	}
+	100% {
+		background-position: -135% 0%;
+	}
 }
