+++ const/contracts.ts
@@ -1,17 +1,28 @@
+import client from "@/lib/client";
 /** Replace the values below with the addresses of your smart contracts. */
 
 // 1. Set up the network your smart contracts are deployed to.
 // First, import the chain from the package, then set the NETWORK variable to the chain.
-import { Mumbai } from "@thirdweb-dev/chains";
-export const NETWORK = Mumbai;
+import { getContract } from "thirdweb";
+import { sepolia } from "thirdweb/chains";
+export const NETWORK = sepolia;
 
 // 2. The address of the marketplace V3 smart contract.
 // Deploy your own: https://thirdweb.com/thirdweb.eth/MarketplaceV3
-export const MARKETPLACE_ADDRESS = "0x83c57ec0dF015eef8401fFb7CB7f66CfA8b6Ff55";
+const MARKETPLACE_ADDRESS = "0x38ab4489E479c9266471bbe8C3794CB30EA11F20";
+export const MARKETPLACE = getContract({
+	address: MARKETPLACE_ADDRESS,
+	client,
+	chain: NETWORK,
+});
 
 // 3. The address of your NFT collection smart contract.
-export const NFT_COLLECTION_ADDRESS =
-  "0x67b40fC017863743C520cDAec1B367BE1FD721f0";
+const NFT_COLLECTION_ADDRESS = "0x72a6eb347D86Bb5DE9c3c6a3DFAb6f2eff80F3C9";
+export const NFT_COLLECTION = getContract({
+	address: NFT_COLLECTION_ADDRESS,
+	client,
+	chain: NETWORK,
+});
 
 // (Optional) Set up the URL of where users can view transactions on
 // For example, below, we use Mumbai.polygonscan to view transactions on the Mumbai testnet.
