+++ next.config.js
@@ -1,7 +1,15 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  reactStrictMode: true,
-  basePath: "",
+	reactStrictMode: true,
+	basePath: "",
+	images: {
+		remotePatterns: [
+			{
+				protocol: "https",
+				hostname: "**",
+			},
+		],
+	},
 };
 
 module.exports = nextConfig;
