+++ package.json
@@ -11,22 +11,23 @@
     "deploy": "next build && next export && npx thirdweb@latest upload out"
   },
   "dependencies": {
-    "@thirdweb-dev/chains": "^0",
-    "@thirdweb-dev/react": "^4",
-    "@thirdweb-dev/sdk": "^4",
-    "ethers": "^5",
-    "next": "^13",
-    "nextjs-progressbar": "^0.0.16",
-    "react": "^18.2",
-    "react-dom": "^18.2",
+    "classnames": "^2.5.1",
+    "next": "^14",
+    "pino-pretty": "^11.0.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-hook-form": "^7.43.0",
-    "react-hot-toast": "^2.4.0"
+    "react-hot-toast": "^2.4.0",
+    "thirdweb": "^5.16.0-nightly-5b0c37a0d9b7ba3d7f38647bb41463e91cc91a49-20240504030706"
   },
   "devDependencies": {
     "@types/node": "^18.11.11",
     "@types/react": "^18",
+    "autoprefixer": "^10.4.19",
     "eslint": "^8.29.0",
-    "eslint-config-next": "^13",
+    "eslint-config-next": "^14.2.3",
+    "postcss": "^8.4.38",
+    "tailwindcss": "^3.4.3",
     "typescript": "^4.9.4"
   }
 }
